requirements.txt
streamlit>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
